+++ src/app/basement/index.js
@@ -10,6 +10,7 @@ require('./directives/keystroke_directive');
 require('./filters/timesince_filter');
 require('./filters/timeuntil_filter');
 require('./filters/bytes_filter');
+require('./multi_select/multi_select');
 require('./nothing/nothing');
 require('./view/top_bar');
 require('./view/view_buttons');