+++ src/app/entry/entry_controller.js
@@ -1,9 +1,10 @@
 "use strict";
 
 import app from "../app";
+require("./entry.css");
 
-function EntryController($scope, $state, $stateParams, Entry, UndoService, Codekit, API, AuthService, Pagination, Search, localStorageService,
-                        $translate) {
+function EntryController($scope, $state, Entry, UndoService, Codekit, API, AuthService, Pagination, Search,
+  localStorageService, $translate) {
 
   function constructor() {
     $scope.undoService = UndoService;
@@ -19,17 +20,20 @@ function EntryController($scope, $state, $stateParams, Entry, UndoService, Codek
     $scope.entries = [];
 
     $translate([
-      'NO_PAGES', 'NO_POSTS', 'DRAFT', 'PUBLISHED',
-      'PIN', 'UNPIN', "ENABLE_COMMENTS", 'DISABLE_COMMENTS'
+      'NO_PAGES', 'NO_POSTS', 'REMOVE_SELECTED_PAGES_PROMPT', 'REMOVE_SELECTED_POSTS_PROMPT', 'DRAFT',
+      'HIDE_FROM_PUBLIC', 'PUBLISHED', 'PIN_FRONT_PAGE', 'UNPIN_FRONT_PAGE', "ALLOW_COMMENTING", 'DISABLE_COMMENTING'
     ]).then(function (translation) {
       if ($scope.isPageView) {
         $scope.nothingText = translation.NO_PAGES;
+        $scope.removeSelectedPrompt = translation.REMOVE_SELECTED_PAGES_PROMPT;
       } else {
         $scope.nothingText = translation.NO_POSTS;
+        $scope.removeSelectedPrompt = translation.REMOVE_SELECTED_POSTS_PROMPT;
       }
 
       $scope.actions = [{
         label: translation.DRAFT,
+        tooltip: translation.HIDE_FROM_PUBLIC,
         icon: "pencil",
         property: "status",
         value: 0
@@ -37,24 +41,24 @@ function EntryController($scope, $state, $stateParams, Entry, UndoService, Codek
         label: translation.PUBLISHED,
         icon: "globe",
         property: "status",
-        value: 1,
+        value: 1
       }, {
-        label: translation.PIN,
+        label: translation.PIN_FRONT_PAGE,
         icon: "star",
         property: "featured",
         value: true
       }, {
-        label: translation.UNPIN,
+        label: translation.UNPIN_FRONT_PAGE,
         icon: "star-o",
         property: "featured",
         value: false
       }, {
-        label: translation.ENABLE_COMMENTS,
+        label: translation.ALLOW_COMMENTING,
         icon: "comments",
         property: "comment_enabled",
         value: true
       }, {
-        label: translation.DISABLE_COMMENTS,
+        label: translation.DISABLE_COMMENTING,
         icon: "ban",
         property: "comment_enabled",
         value: false
@@ -78,48 +82,6 @@ function EntryController($scope, $state, $stateParams, Entry, UndoService, Codek
       }
     );
   }
-
-  /**
-   * @desc set property of selected entries
-   *
-   * @param {string} key
-   * @param {boolean|number} value
-   */
-  $scope.setProperty = function(key, value) {
-    angular.forEach($scope.entries, function(entry) {
-      if (entry.isSelected) {
-        entry.setProperty(key, value);
-      }
-    });
-  };
-
-  /**
-   * @desc Remove selected entries
-   */
-  $scope.removeSelected = function() {
-
-    if (confirm($translate.instant('REMOVE_SELECTED_ENTRY_PROMPT')) === true) {
-      angular.forEach($scope.entries, function(entry) {
-        if (entry.isSelected) {
-          entry.remove();
-        }
-      });
-    } else {
-      return;
-    }
-  };
-
-  /**
-   * @desc Count selected entries
-   */
-  $scope.countSelected = function() {
-    $scope.selectCount = 0;
-    angular.forEach($scope.entries, function(entry) {
-      if (entry.isSelected) {
-        $scope.selectCount++;
-      }
-    });
-  };
 
   /**
    * @desc Load more function for controller
@@ -166,7 +128,6 @@ app.controller("EntryController", EntryController);
 EntryController.$inject = [
   "$scope",
   "$state",
-  "$stateParams",
   "Entry",
   "UndoService",
   "Codekit",