+++ CHANGELOG.md
@@ -9,6 +9,10 @@
 - Resolved #1003: Add a method to cancel blog's current subscription.
 - Resolved #1002: Add upgrade plans to blog settings view.
 - Resolved #1001: Add transactions list to blog settings view.
+- Resolved #1000: Add a method to multi select and delete selected items.
+- Fixed #1005: Cancel item deletion timeout on undo.
+- Resolved #1008: Filter comments based on post/page.
+- Resolved #1009: Update new navbar color and remove it's animation.
 
 
 2018.10.1